Do not unset config @ cluster.destroy
Destroying the cluster doesn't have to destroy the configs that live in
the dev env, and I found that if there is a problem with the cluster
that I have to delete and recreate it, I often forget to re-set the
configs. This way, the configs stay in place, and I save time and
frustration.

Signed-off-by: Blaine Gardner <blaine.gardner@suse.com>